Create nlp_twitter_chatgpt_q12023.ipynb
This Python project is a continuation, revision and extension of the recent data science study of 500k tweets about ChatGPT in Q1'23. Our objective is to unveil public opinion and ChatGPT role in shaping the future of AI-powered LLMs.  This includes text processing, sentiment analysis, identification of influencers, and advanced visualization of NLP word clouds, Ngram counts, topic modelling LDA, and stock impacts.